catalog: Move RBAC structs to the sql crate
This commit moves RBAC related structs from the mz_adapter::catalog
crate/module to the mz_sql::catalog crate/module. This will help in the
effort to extract catalog::storage into its own crate.

Works towards resolving #20953
