+++ src/adapter/src/catalog/storage/stash.rs
@@ -15,7 +15,7 @@ use mz_ore::now::EpochMillis;
 use mz_proto::ProtoType;
 use mz_repr::adt::mz_acl_item::AclMode;
 use mz_repr::role_id::RoleId;
-use mz_sql::catalog::{RoleAttributes, SystemObjectType};
+use mz_sql::catalog::{RoleAttributes, RoleMembership, SystemObjectType};
 use mz_sql::names::{
     DatabaseId, ObjectId, ResolvedDatabaseSpecifier, SchemaId, SchemaSpecifier, PUBLIC_ROLE_NAME,
 };
@@ -26,9 +26,9 @@ use mz_storage_client::types::sources::Timeline;
 use crate::catalog::builtin::{MZ_SUPPORT_ROLE, MZ_SYSTEM_ROLE};
 use crate::catalog::object_type_to_audit_object_type;
 use crate::catalog::storage::{
-    BootstrapArgs, DefaultPrivilegesKey, DefaultPrivilegesValue, RoleMembership,
-    SystemPrivilegesKey, SystemPrivilegesValue, AUDIT_LOG_ID_ALLOC_KEY, DATABASE_ID_ALLOC_KEY,
-    MZ_SUPPORT_ROLE_ID, MZ_SYSTEM_ROLE_ID, SCHEMA_ID_ALLOC_KEY, STORAGE_USAGE_ID_ALLOC_KEY,
+    BootstrapArgs, DefaultPrivilegesKey, DefaultPrivilegesValue, SystemPrivilegesKey,
+    SystemPrivilegesValue, AUDIT_LOG_ID_ALLOC_KEY, DATABASE_ID_ALLOC_KEY, MZ_SUPPORT_ROLE_ID,
+    MZ_SYSTEM_ROLE_ID, SCHEMA_ID_ALLOC_KEY, STORAGE_USAGE_ID_ALLOC_KEY,
     SYSTEM_CLUSTER_ID_ALLOC_KEY, SYSTEM_REPLICA_ID_ALLOC_KEY, USER_CLUSTER_ID_ALLOC_KEY,
     USER_REPLICA_ID_ALLOC_KEY, USER_ROLE_ID_ALLOC_KEY,
 };
